Fixed game null image bug & improved end-all-plays
A bug where a 500 internal error occurred when trying to access the URL
attribute of an image via the game object through the Game serializers
used by the views was fixed.
Now it doesn't try to access the URL attribute from the image directly
but first, it checks if the game has an image, and case it's None a
null response is returned, and no attempt to access the URL attribute
of the inexistent image is made.

Also, the end_all_plays URL is being renamed to end-all-plays as the
Google Developer Guides recommends
developers.google.com/search/docs/crawling-indexing/url-structure

A final improvement to the swagger in the end-all-plays endpoint is
made by updating that no request body is required to successfully use
the endpoint