Update regex validations, add pylint and fix redis
Some regex validations were updated to match the new requirements.
Also, some exception handlers were implemented to avoid errors, and
improve feedback.

Pylint was added to the project, and some fixes were made to the code
to match the pylint requirements. #72

The failing requests due to the redis server being down were fixed by
adding a try-except block to handle the exception and return a proper
response to the user even if the message wasn't successfully sent.
fixes #80